Update main.py
Addressing issue with how the paths are being formed in the os.walk function. os.walk generates the file paths relative to the directory you passed to it, if you pass a directory name without any path it will look for that directory in the current working directory and create paths relative to that directory. 
- updated function which should correctly build the file paths
- looks for ethereum, bsc, solana, polkadot directories in the current working directory (and adds all .sol and .rs files to the listbox, displaying their paths relative to their respective directories